refactor: extract subentry methods to SubentryOperations
Moves all subentry-related methods from main.py to subentry.py:
- SubentryMetadata dataclass and _sanitize_subentry_identifier helper
- attach_subentry_manager, _default_subentry_key
- async_wait_subentry_visibility_updates
- _build_core_subentry_definitions, _schedule_core_subentry_repair
- _cancel_pending_subentry_repair, _refresh_subentry_index
- _group_snapshot_by_subentry, _store_subentry_snapshots
- _resolve_subentry_key_for_feature, get_subentry_key_for_feature
- get_subentry_metadata, stable_subentry_identifier
- get_subentry_snapshot, is_device_visible_in_subentry
- get_device_location_data_for_subentry, get_device_last_seen_for_subentry

Also removes unused imports from main.py.

main.py: 5,313 lines (from 8,342 = 36% reduction)
subentry.py: 919 lines
